`.setActivity` is the new method used to designate the activity that the bot is performing currently on Discord.
If the user has `HUBOT_DISCORD_STATUS_MSG` as an environment variable,
this will be passed and set as the bot's status with whatever the variable
is set to. This will show up as the bot is "Playing" whatever the string
is.

Example:
`HUBOT_DISCORD_STATUS_MSG="DOOM 2016 OST"`

```
HUBOT
Playing DOOM OST
```